+++ api/product/src/main/java/store/product/ProductIn.java
@@ -6,7 +6,7 @@ import lombok.experimental.Accessors;
 @Builder @Accessors(fluent = true)
 public record ProductIn(
     String name,
-    Float price,
+    Double price,
     String unit
 ) {
